requirements.txt
Flask==2.3.3
pytesseract==0.3.10
opencv-python-headless==4.8.0.76
Pillow==10.0.0
numpy==1.24.3
python-dateutil==2.8.2
Werkzeug==2.3.7
gunicorn==21.2.0
